% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\title{Fill a Template}
\usage{
fill(template, spec, values, value_col)
}
\arguments{
\item{template}{A character vector, each element of which is a line of LaTeX code.}

\item{spec}{A list describing the relationship between values and template. See Details.}

\item{values}{A tidy dataframe containing the values that will populate the template.}

\item{value_col}{The name of the column in \code{values} that will}
}
\description{
Fills in a TeX template with values according to spec
}
